backend/services/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
matplotlib==3.8.3
ipython
